Move overlap-correlation into service
